crypto-kalman-reversion/scripts/requirements.txt
alpaca==1.0.0
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
python-dotenv==1.1.0
statsmodels==0.14.4
ta==0.11.0
tqdm==4.67.1
